Move handlers out of hpx_init.cpp
